Add --gzip-block-factor option, to prevent runaway memory usage
Gzip files can't necessarily be resumed from arbitrary positions. Instead, some or all blocks require a 32KB dictionary to decode them. If such a dict is needed for some block, it is included in the index file.

By default, we cut the gzip file at boundaries of 32 dicts (uncompressed) so that the index file will be at most about 3% of the uncompressed file size. However, for very large files, that can rapidly eat up all memory. If this is a problem, the caller can set a larger boundary multiple, eg: --gzip-block-factor=1024 .